meddata/requirements.txt
sentence-transformers
umap-learn
plotly
numpy
scikit-learn
pandas
